decoding of tlv and incoming data from client to server

+++ server/pages/models/ContentValueModel.go
@@ -1,6 +1,6 @@
 package models
 
 type ContentValueModel struct {
-	Id 		int
-	Value 	interface{}
+	Id    uint16
+	Value interface{}
 }
